Get text tests passing
