+++ python/cudf/cudf/core/column/lists.py
@@ -8,7 +8,6 @@ from typing import TYPE_CHECKING, Any, Literal, cast
 
 import pandas as pd
 import pyarrow as pa
-from typing_extensions import Self
 
 import pylibcudf as plc
 
@@ -32,6 +31,7 @@ from cudf.utils.utils import _is_null_host_scalar
 if TYPE_CHECKING:
     from collections.abc import Sequence
 
+    import numpy as np
     from typing_extensions import Self
 
     from cudf._typing import ColumnBinaryOperand, ColumnLike, Dtype
@@ -512,3 +512,39 @@ class ListColumn(ColumnBase):
             plc.strings.combine.OutputIfEmptyList.NULL_ELEMENT,
         )
         return type(self).from_pylibcudf(plc_column)  # type: ignore[return-value]
+
+    @acquire_spill_lock()
+    def minhash_ngrams(
+        self,
+        width: int,
+        seed: np.uint32,
+        a: NumericalColumn,
+        b: NumericalColumn,
+    ) -> Self:
+        return type(self).from_pylibcudf(  # type: ignore[return-value]
+            plc.nvtext.minhash.minhash_ngrams(
+                self.to_pylibcudf(mode="read"),
+                width,
+                seed,
+                a.to_pylibcudf(mode="read"),
+                b.to_pylibcudf(mode="read"),
+            )
+        )
+
+    @acquire_spill_lock()
+    def minhash64_ngrams(
+        self,
+        width: int,
+        seed: np.uint64,
+        a: NumericalColumn,
+        b: NumericalColumn,
+    ) -> Self:
+        return type(self).from_pylibcudf(  # type: ignore[return-value]
+            plc.nvtext.minhash.minhash64_ngrams(
+                self.to_pylibcudf(mode="read"),
+                width,
+                seed,
+                a.to_pylibcudf(mode="read"),
+                b.to_pylibcudf(mode="read"),
+            )
+        )
